fix: friends refactor
